feat: agregar interfaz IPermiso y actualizar usuario para incluir permisos
