+++ src/interfaces/index.ts
@@ -46,6 +46,7 @@ export * from './twilio-templates';
 export * from './ubicacion';
 export * from './visita-ubicacion';
 export * from './usuario';
+export * from './permiso';
 export * from './destinatario-asistencia';
 export * from './emergencias';
 export * from './personal-salud';
