+++ src/interfaces/usuario.ts
@@ -1,9 +1,5 @@
-import { IActivo } from './activo';
 import { ICliente } from './cliente';
-import { IDispositivoAlarma } from './dispositivo-alarma';
-import { IGrupo } from './grupo';
-import { ILuminaria } from './luminaria';
-import { IRol } from './rol';
+import { IPermiso } from './permiso';
 
 export type Rol =
   | 'Administrador'
@@ -61,80 +57,11 @@ export interface IClaveUsuarioAlarma {
   clave?: string;
 }
 
-export interface IVencimientoPermisoUsuario {
-  // Venicimiento de un permiso con sus respectivas opciones
-  fechaVencimiento?: string;
-  eliminarPermiso?: boolean;
-  desactivarUsuario?: boolean;
-  eliminarUsuario?: boolean;
-}
-
-export interface ICredencialesSeguridad {
-  usuario?: string;
-  claveEncriptada?: string;
-}
-
-//Par actualizar credenciales de seguridad vía PUT /:id/credencialesSeguridad, se recibe este DTO con los datos en texto plano
-// (la gestion-api-gestion se encarga de cifrar la clave antes de persistir)
-export interface IUpdateCredencialesSeguridad {
-  idCliente: string;
-  usuario: string;
-  clave: string;
-  // Índice del permiso en el array, para distinguir múltiples permisos con el mismo idCliente
-  permisoIndex?: number;
-}
-
-export type Nivel = 'Cliente' | 'Grupo' | 'Entidad';
-export type TipoEntidadPermiso =
-  | 'Activo'
-  | 'Vehículo'
-  | 'Colectivo'
-  | 'Luminaria'
-  | 'Alarma'
-  | 'Cámara';
-
-export interface IPermiso {
-  nivel?: Nivel;
-  // Para nivel Cliente
-  idCliente?: string;
-  includeChildren?: boolean;
-  idsAncestros?: string[];
-  // Para nivel Grupo
-  idsGrupos?: string[];
-  // Para nivel Entidad
-  tipoEntidad?: TipoEntidadPermiso;
-  idsEntidades?: string[];
-  //
-  activo?: boolean; // Si el permiso está activo o no
-  vencimiento?: IVencimientoPermisoUsuario;
-  modulos?: IModulos;
-
-  /**
-   * @deprecated El campo 'roles' está en desuso. Se eliminará en futuras versiones. Los reemplaza "idsRoles" y el virtual "Rols"
-   */
-  roles?: Rol[];
-
-  idsRoles?: string[]; // IDs de los roles asignados al permiso
-
-  credencialesSeguridad?: ICredencialesSeguridad; // Credenciales del sistema de seguridad (AES encriptadas)
-  tieneCredencialesSeguridad?: boolean; // Virtual: true si el permiso tiene credenciales configuradas
-
-  // Virtual
-  cliente?: ICliente;
-  ancestros?: ICliente[];
-  grupos?: IGrupo[];
-  activos?: IActivo[]; // Entidades pobladas según el tipoEntidad
-  luminarias?: ILuminaria[]; // Entidades pobladas según el tipoEntidad
-  alarmas?: IDispositivoAlarma[]; // Entidades pobladas según el tipoEntidad
-  rols?: IRol[]; // Roles poblados según idsRoles
-}
-
 export interface IUsuario {
   _id?: string;
   identificacionInterna?: string;
   idCliente?: string;
   idsAncestros?: string[];
-  permisos?: IPermiso[];
   //
   idExterno?: string;
   fechaCreacion?: string;
@@ -142,9 +69,12 @@ export interface IUsuario {
   hash?: string;
   datosPersonales?: IDatosPersonales;
   config?: IConfigUsuario;
-  // Populate
+  // Populate / Virtual
   cliente?: ICliente;
   ancestros?: ICliente[];
+  // Permisos del usuario. Ya no es un array embebido: es un virtual poblado
+  // desde la colección Permiso por match de nombreUsuario (Permiso.nombreUsuario === Usuario.usuario).
+  permisos?: IPermiso[];
 }
 
 type OmitirCreate = '_id' | 'cliente' | 'fechaCreacion';
